28996: Dev branch for De-Lucene work pending patches
   29004: Evaluator runs in read-only txn
   29006: Additional PermissionCheckedCollection.create method
          - Use an existing collection's permission check data (cut-off, etc) to wrap a new collection
   29007:
          CopyService and CheckOutCheckInService refactors to remove Lucene
          
          CopyService:
          
          Removed cm:source property from cm:copiedfrom aspect and replaced with a cm:original association.
          Added CQ-based APIs to query for copies
          Added APIs to support bi-directional walking of copy association
          Fixed sundry uses of cm:copiedfrom esp. all uses related to cm:workingcopy
          
          CheckOutCheckInService:
          
          Check-out now creates a source aspect cm:checkedOut with 1:1 relationship to cm:workingcopy via cm:workingcopylink
          Removed explicit use of cm:workingcopy aspect and replaced it with calls to COCI API
          
   29083: Audit tests fail when indexing is turned off.
          Also removed a getReader() call during rule evaluation, leading to a 'sub-action' read being recorded.
   29113: NodeDAO.getNodesWithAspects supports paging
   29135: Removed unused patch queries
   29139: Basic patch (still terminates with error) to upgrade cm:copiedfrom and cm:workingcopy
   29157: Tested patch for cm:copiedfrom and cm:workingcopy aspects


git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@29159 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/repo/lock/LockServiceImpl.java
@@ -56,8 +56,6 @@ import org.alfresco.service.cmr.repository.StoreRef;
 import org.alfresco.service.cmr.search.ResultSet;
 import org.alfresco.service.cmr.search.SearchService;
 import org.alfresco.service.cmr.security.AuthenticationService;
-import org.alfresco.service.cmr.security.OwnableService;
-import org.alfresco.service.namespace.NamespaceService;
 import org.alfresco.service.namespace.QName;
 
 /**
@@ -73,100 +71,36 @@ public class LockServiceImpl implements LockService,
                                         VersionServicePolicies.BeforeCreateVersionPolicy,
                                         VersionServicePolicies.OnCreateVersionPolicy
 {
-    /**
-     * The node service
-     */
-    private NodeService nodeService;
-    
-    /**
-     * The tenant service
-     */
-    private TenantService tenantService;
-
-    /**
-     * The policy component
-     */
-    private PolicyComponent policyComponent;
-    
     /** Key to the nodes ref's to ignore when checking for locks */
     private static final String KEY_IGNORE_NODES = "lockService.ignoreNodes";
     
-    /**
-     * The authentication service
-     */
+    private NodeService nodeService;
+    private TenantService tenantService;
     private AuthenticationService authenticationService;
-
-    /**
-     * The ownable service
-     * 
-     */
-    private OwnableService ownableService;
-    
-    /**
-     * The search service
-     */
     private SearchService searchService;
 
-    /**
-     * Set the node service
-     * 
-     * @param nodeService
-     *            the node service
-     */
+    private PolicyComponent policyComponent;
+    
     public void setNodeService(NodeService nodeService)
     {
         this.nodeService = nodeService;
     }
  
-    /**
-     * Set the tenant service
-     * 
-     * @param tenantService
-     *            the tenant service
-     */   
     public void setTenantService(TenantService tenantService)
     {
         this.tenantService = tenantService;
     }
 
-    /**
-     * Sets the policy component
-     * 
-     * @param policyComponent
-     *            the policy componentO
-     */
     public void setPolicyComponent(PolicyComponent policyComponent)
     {
         this.policyComponent = policyComponent;
     }
 
-    /**
-     * Sets the authentication service
-     * 
-     * @param authenticationService
-     *            the authentication service
-     */
     public void setAuthenticationService(AuthenticationService authenticationService)
     {
         this.authenticationService = authenticationService;
     }
 
-    /**
-     * Sets the ownable service
-     * 
-     * @param ownableService
-     *            the ownable service
-     */
-    public void setOwnableService(OwnableService ownableService)
-    {
-        this.ownableService = ownableService;
-    }
-    
-    /**
-     * Set the search service
-     * 
-     * @param searchService     the search service
-     */
     public void setSearchService(SearchService searchService)
     {
         this.searchService = searchService;
@@ -353,19 +287,25 @@ public class LockServiceImpl implements LockService,
     public void unlock(NodeRef nodeRef) throws UnableToReleaseLockException
     {
         nodeRef = tenantService.getName(nodeRef);
-        
-        // Check for lock aspect
-        checkForLockApsect(nodeRef);
-        
-        addToIgnoreSet(nodeRef);
-        try
+        if (!nodeService.hasAspect(nodeRef, ContentModel.ASPECT_LOCKABLE))
         {
-            // Clear the lock
-            this.nodeService.removeAspect(nodeRef, ContentModel.ASPECT_LOCKABLE);
+            // Nothing to unlock
         }
-        finally
+        else
         {
-            removeFromIgnoreSet(nodeRef);
+            // Check for lock aspect
+            checkForLockApsect(nodeRef);
+            
+            addToIgnoreSet(nodeRef);
+            try
+            {
+                // Clear the lock
+                this.nodeService.removeAspect(nodeRef, ContentModel.ASPECT_LOCKABLE);
+            }
+            finally
+            {
+                removeFromIgnoreSet(nodeRef);
+            }
         }
     }
